src/backend/servico_contas/requirements.txt
# Framework web e servidor ASGI
fastapi
uvicorn[standard]

# ORM e Migrações de Banco de Dados
sqlalchemy[asyncio]
alembic
asyncpg

# Validação de dados
pydantic[email]
pydantic
python-multipart


# Dependências de Segurança
passlib[bcrypt]
python-jose[cryptography]

# Adicione esta linha para fixar a versão do bcrypt
bcrypt==3.2.0